refactor: move abstracted tab logic to origin file
Firefox's dev tools showed the warning that all files should have a default export of a function, but due to the nature of their purpose, these two files didn't have such exports. Also, the structure was confusing (declaring a type in the file that didn't end in types.ts) and unnecessarily 'counterintuitive' (two files with a not very clear motive for separating).